chore: PR 리뷰 기반 스타일링 및 구조 개선
- theme 변수로 색상 상수화 (e.g. #FD906F → theme.color.orangeChat)
- 불필요한 rem → px 혼용 제거 및 일관성 유지
- 불필요하거나 중복되는 스타일링 제거
- 라우트 명명 규칙 통일 (/friend/dm → /friend/chatroom)
- 컴포넌트 분리로 의존성 감소 및 가독성 향상
