+++ src/App.tsx
@@ -77,7 +77,7 @@ const App = () => {
         { path: PATH.HISTORY, element: <HistoryPage /> },
         { path: PATH.PROFILE, element: <ProfilePage /> },
         { path: PATH.FRIEND, element: <FriendPage /> },
-        { path: PATH.FRIEND_DM, element: <FriendChatRoomPage /> },
+        { path: PATH.FRIEND_CHATROOM, element: <FriendChatRoomPage /> },
       ],
     },
   ];
