+++ src/pages/FriendPage/components/friend-list/styles.css.ts
@@ -1,11 +1,12 @@
 import { style } from '@vanilla-extract/css';
 
 import { rem } from '@/styles';
+import { theme } from '@/styles/theme.css';
 
 export const friendList = style({
   width: rem(200),
   height: '100%',
-  color: 'white',
+  color: theme.color.white,
   paddingTop: rem(50),
   overflowY: 'auto',
   overflowX: 'hidden',
@@ -25,14 +26,14 @@ export const item = style({
   gap: rem(10),
   cursor: 'pointer',
   padding: `${rem(6)} ${rem(1)} ${rem(6)} ${rem(9)}`,
-  borderBottom: `${rem(1.5)} solid #ffffff`,
+  borderBottom: `${rem(1.5)} solid ${theme.color.white}`,
   fontSize: rem(22),
   whiteSpace: 'nowrap',
   overflow: 'hidden',
   textOverflow: 'ellipsis',
 
   ':hover': {
-    backgroundColor: '#333',
+    backgroundColor: theme.color.grayHover,
   },
 });
 
@@ -46,5 +47,5 @@ export const avatar = style({
 export const divider = style({
   width: rem(200),
   height: rem(1),
-  backgroundColor: '#ffffff',
+  backgroundColor: theme.color.white,
 });
